Puts the createDirectMessage on the proper package.
