+++ core/src/main/kotlin/chat/rocket/core/internal/rest/Message.kt
@@ -4,7 +4,6 @@ import chat.rocket.common.model.BaseResult
 import chat.rocket.common.model.RoomType
 import chat.rocket.core.RocketChatClient
 import chat.rocket.core.internal.RestResult
-import chat.rocket.core.internal.model.CreateDirectMessagePayload
 import chat.rocket.core.internal.model.DeletePayload
 import chat.rocket.core.internal.model.MessageReportPayload
 import chat.rocket.core.internal.model.PostMessagePayload
@@ -13,7 +12,6 @@ import chat.rocket.core.internal.model.SendMessageBody
 import chat.rocket.core.internal.model.SendMessagePayload
 import chat.rocket.core.model.DeleteResult
 import chat.rocket.core.model.Message
-import chat.rocket.core.model.NewDirectMessageResult
 import chat.rocket.core.model.PagedResult
 import chat.rocket.core.model.ReadReceipt
 import chat.rocket.core.model.attachment.Attachment
@@ -398,25 +396,3 @@ suspend fun RocketChatClient.reportMessage(
 
     return@withContext handleRestCall<BaseResult>(request, BaseResult::class.java).success
 }
-
-/**
- * Create a direct message session with another user.
- *
- * @param username The username of the user to create a session with.
- * @return The updated Message object.
- */
-suspend fun RocketChatClient.createDirectMessage(username: String): NewDirectMessageResult =
-    withContext(CommonPool) {
-        val payload = CreateDirectMessagePayload(username = username)
-        val adapter = moshi.adapter(CreateDirectMessagePayload::class.java)
-        val payloadBody = adapter.toJson(payload)
-
-        val body = RequestBody.create(MEDIA_TYPE_JSON, payloadBody)
-
-        val url = requestUrl(restUrl, "im.create").build()
-        val request = requestBuilderForAuthenticatedMethods(url).post(body).build()
-
-        val type = Types.newParameterizedType(RestResult::class.java, NewDirectMessageResult::class.java)
-
-        return@withContext handleRestCall<RestResult<NewDirectMessageResult>>(request, type).result()
-    }
