+++ core/src/test/kotlin/chat/rocket/core/internal/rest/MessagesTest.kt
@@ -2,7 +2,6 @@ package chat.rocket.core.internal.rest
 
 import chat.rocket.common.RocketChatException
 import chat.rocket.common.model.Token
-import chat.rocket.common.model.roomTypeOf
 import chat.rocket.common.util.PlatformLogger
 import chat.rocket.core.RocketChatClient
 import chat.rocket.core.TokenRepository
@@ -253,24 +252,6 @@ class MessagesTest {
         runBlocking {
             val result = sut.reportMessage("FCHGvHLyanhbaWjpxWz", "something untoward was said")
             assertThat(result, isEqualTo(true))
-        }
-    }
-
-    @Test
-    fun `createDirectMessage() should return true and yield no exceptions`() {
-        mockServer.expect()
-            .post()
-            .withPath("/api/v1/im.create")
-            .andReturn(200, CREATE_DM_OK)
-            .once()
-
-        runBlocking {
-            val result = sut.createDirectMessage(username = "rocket.cat")
-            assertThat(result.id, isEqualTo("Lymsiu4Mn6xjTAan4RtMDEYc28fQ5aHpf4"))
-            assertThat(result.type, isEqualTo(roomTypeOf("d")))
-            assertThat(result.usernames.size, isEqualTo(2))
-            assertThat(result.usernames[0], isEqualTo("rocket.cat"))
-            assertThat(result.usernames[1], isEqualTo("user.test"))
         }
     }
